refactor(player): use stable queue entries
