+++ Sources/Kaset/Services/Player/PlayerQueueModels.swift
@@ -1,0 +1,15 @@
+import Foundation
+
+// MARK: - QueueEntry
+
+struct QueueEntry: Identifiable, Hashable {
+    let id: UUID
+    let song: Song
+}
+
+// MARK: - QueueState
+
+struct QueueState {
+    let entries: [QueueEntry]
+    let currentIndex: Int
+}
